+++ test/sveltekit/templates.test.ts
@@ -39,6 +39,10 @@ describe('getClientHooksTemplate', () => {
 
         // If you don't want to use Session Replay, just remove the line below:
         integrations: [replayIntegration()],
+
+        // Enable sending user PII (Personally Identifiable Information)
+        // https://docs.sentry.io/platforms/javascript/guides/sveltekit/configuration/options/#sendDefaultPii
+        sendDefaultPii: true,
       });
 
       // If you have a custom error handler, pass it to \`handleErrorWithSentry\`
@@ -72,6 +76,10 @@ describe('getClientHooksTemplate', () => {
 
         // If you don't want to use Session Replay, just remove the line below:
         integrations: [replayIntegration()],
+
+        // Enable sending user PII (Personally Identifiable Information)
+        // https://docs.sentry.io/platforms/javascript/guides/sveltekit/configuration/options/#sendDefaultPii
+        sendDefaultPii: true,
       });
 
       // If you have a custom error handler, pass it to \`handleErrorWithSentry\`
@@ -98,6 +106,10 @@ describe('getClientHooksTemplate', () => {
 
 
 
+
+        // Enable sending user PII (Personally Identifiable Information)
+        // https://docs.sentry.io/platforms/javascript/guides/sveltekit/configuration/options/#sendDefaultPii
+        sendDefaultPii: true,
       });
 
       // If you have a custom error handler, pass it to \`handleErrorWithSentry\`
@@ -124,6 +136,10 @@ describe('getClientHooksTemplate', () => {
         enableLogs: true,
 
 
+
+        // Enable sending user PII (Personally Identifiable Information)
+        // https://docs.sentry.io/platforms/javascript/guides/sveltekit/configuration/options/#sendDefaultPii
+        sendDefaultPii: true,
       });
 
       // If you have a custom error handler, pass it to \`handleErrorWithSentry\`
@@ -158,6 +174,11 @@ describe('getServerHooksTemplate', () => {
         // Enable logs to be sent to Sentry
         enableLogs: true,
 
+
+        // Enable sending user PII (Personally Identifiable Information)
+        // https://docs.sentry.io/platforms/javascript/guides/sveltekit/configuration/options/#sendDefaultPii
+        sendDefaultPii: true,
+
         // uncomment the line below to enable Spotlight (https://spotlightjs.com)
         // spotlight: import.meta.env.DEV,
       });
@@ -190,6 +211,11 @@ describe('getServerHooksTemplate', () => {
       Sentry.init({
         dsn: 'https://sentry.io/123',
 
+
+
+        // Enable sending user PII (Personally Identifiable Information)
+        // https://docs.sentry.io/platforms/javascript/guides/sveltekit/configuration/options/#sendDefaultPii
+        sendDefaultPii: true,
 
         // uncomment the line below to enable Spotlight (https://spotlightjs.com)
         // spotlight: import.meta.env.DEV,
@@ -225,6 +251,11 @@ describe('getServerHooksTemplate', () => {
 
         // Enable logs to be sent to Sentry
         enableLogs: true,
+
+
+        // Enable sending user PII (Personally Identifiable Information)
+        // https://docs.sentry.io/platforms/javascript/guides/sveltekit/configuration/options/#sendDefaultPii
+        sendDefaultPii: true,
 
         // uncomment the line below to enable Spotlight (https://spotlightjs.com)
         // spotlight: import.meta.env.DEV,